+++ torch_xla/experimental/xla_sharded_tensor.py
@@ -4,7 +4,7 @@ import warnings
 
 warnings.warn(
     "Importing from `torch_xla.experimental.xla_sharded_tensor` will be deprecated "
-    "after 2.2 release. Please use `torch_xla.experimental.spmd` "
+    "after 2.2 release. Please use `torch_xla.distributed.spmd` "
     "instead.", DeprecationWarning, 2)
 
-from .spmd.xla_sharded_tensor import *
+from torch_xla.distributed.spmd.xla_sharded_tensor import *
