+++ torch_xla/experimental/xla_sharding.py
@@ -4,7 +4,7 @@ import warnings
 
 warnings.warn(
     "Importing from `torch_xla.experimental.xla_sharding` will be deprecated "
-    "after 2.2 release. Please use `torch_xla.experimental.spmd` instead.",
+    "after 2.2 release. Please use `torch_xla.distributed.spmd` instead.",
     DeprecationWarning, 2)
 
-from .spmd.xla_sharding import *
+from torch_xla.distributed.spmd.xla_sharding import *
